reformat notebook 264 and move notebook 266 to separate file
